fix: Disable USR1 on Windows (no signal)
